requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
pydantic>=2.0.0
scipy>=1.11.0
